Moved Editor stuff that was not controllers
