+++ src/Umbraco.Abstractions/Editors/IEditorValidator.cs
